requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.3
django-model-utils==4.2.0
django-storages==1.12.2
djangorestframework==3.12.4
mysqlclient==2.0.3
psycopg2==2.9.1
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
